requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg==3.2.1
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
